Fix for Issue #151
Added a brief sentence to the support footer that points users to the Hubpress.io issue tracker on GitHub. I noted that there is a proper Support section that seems to include a web contact form, but I think activating this would encourage folks to just email instead of using GitHub to request assistance.